refactor: pull extension out of editor file ⬆️
